完成了点击回复后向后端传输json

+++ src/main/java/life/majiang/community/community/controller/PublishController.java
@@ -21,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -43,7 +43,7 @@ public class PublishController {
             @RequestParam(value = "tag",required = false) String tag,
             HttpServletRequest request,
             Model model,
-            @RequestParam(value = "id",required = false) Integer id){
+            @RequestParam(value = "id",required = false) Long id){
         //model能与thymeleaf联系，使前端后端数据交互
         model.addAttribute("title",title);
         model.addAttribute("description",description);
